InboxFragment workaround for #8.
The InboxFragment workaround is that the addToInboxList checks if the page you try to add and if it is the first one it clears the list.

- Return empty list instead of null from RESTSessionStorage.getInboxModelList(), when not in cache.

- Also fixed RESTSessionStorage.getInboxModelList() to handle partial pages (less than the page size).
Previously if the page was not fully populated the cache returned null.

- Also made InboxFragment store it's current page in the RESTSessionStorage,
such that when you come back from another activity and you were for example on page 3 you don't have to click 3 times and scroll trough the list to get back.
The position is preserved.
And restored onResume.

Signed-off-by: vladisla <vladisla@student.matnat.uio.no>