RESTSessionStorage different sessions for different users. (issue #9)
A solution to issue #9.

RESTSessionStorage.setActiveSession(String username)
Allows LoginActivity to initiate the storage session.

LoginActivity also sets a default username (for getInstance), just in case.

RESTSessionStorage.getInstance()
Returns any currenty active session.
If no session is active it makes the default one (from LoginActivity) to default and return that one.
If the default session doesn't exist it is created.