Initial design and implementation for caching of the rest conversations.
After analyzing how the conversations are retrieved, and what lists models precisely are saved,
I have completed the first step in the full implementation.

The caching:
RESTSessionStorage already saves the "headers" of conversations (/api/messageConversations) in the DHIS2 api.
The cache is a list of InboxModels.

I have modified the InboxModel.java to have two aditional lists:
A list of ChatModels to cache the messages downloaded by RESTGetConversation.
And a list of NameAndIdModels to cache the members downloaded by RESTGetConversation.

I also added a method to retrieve the relevant InboxModel by providing the index RESTSessionStorage.getInboxModel(index).
The index is decided by the position of the selected element in the gui in InboxFragment (line 140).

Next I will implement modification fo the cache when REST-Delete/Send-Message.java is executed.

Signed-off-by: vladisla <vladisla@student.matnat.uio.no>