REST messages caching. Send, delete, mark read and updating the lists.
Changes to RESTSessionStorage:
Added switches to indicate if we just sent a message or started a new conversation.
Since the response from the server is not very useful with regards of providing the id's of the message or conversation,
such that for example the app could reconstruct the equivalent to the server object, without having to ask for almost the same data that it just sent while sending the new message/conversation.

So the NewMessageActivity and RestChatActivity.messageSent() set these switches/state indicators.
This scheme is necesery because both activities close after the action and switch to another activity all together.

I have also renamed some of the activities and AsyncTasks, to less confusing names.
For example RESTGetConversation instead of RESTGetMessage, which is misleading.
RestDeleteConversation instead of RESTDeleteMessage. DHIS2 doesn't provide single message deletion as far as I know. And if it did it wouldn't be in the scope of the app.
Also RestMarkConversationRead instead of RestMarkRead.

Also all of the above have been modified to receive the index of the selected inboxModel.
This way they obtain reference to the current conversation inboxModel from RESTSessionStorage.
Thus avoiding searching the cache lists for id's when get/set operations need to be made.

Signed-off-by: vladisla <vladisla@student.matnat.uio.no>