Now updates the number of unread messages when you refresh or open a conversation
This fixes #11. Removed RESTUnreadMessages.java because it's data usage was horrible.
It downloaded all the messages before the inbox loaded to count the number of unread
messages. Now the counting is implemented in the InboxFragment, when the conversations
are got from the server.

Restructured methods in HomeActivity for updating the number of unread messages in the
title.