RESTSessionStorage now merges the duplicate inbox entries with those in cache.
Before when storing the InboxModels for a conversation the duplicates  would just be ignored.
This was less than optimal as no actual updates were possible.

Now their fields are copied over to the old instance in the list.
This saves reinserting the InboxModels into cache.
It also preserves the cached InboxModel.messages and InboxModel.members lists.
Such that simply updating the read value doesn't remove the caches of the conversation messages..etc

I also added the official demo server to the suggested server strings.

Also InboxModels had to be changed.
I implemented the equals method, so the ArrayList can detect equal Conversations. (InboxModels)

Signed-off-by: vladisla <vladisla@student.matnat.uio.no>