Refactored/Abstracted RESTSessionStorage list and page operations.
They are now part of class CacheList<T extends CopyAttributes>.

-New interface to reflect the necessary method copyAttributesFrom().
-InboxModel and InterpretationsModel implement that interface now.

The Interpretations page caching will be pretty much the same as the InboxModel one.
Thus I decided to refactor the classes to abstract and reuse the functionality instead of duplicating code.
This should lead to less duplicate code, better readability and less errors.


Signed-off-by: vladisla <vladisla@student.matnat.uio.no>