Fix for read message status. InboxModel.read and the DHIS2 API.
This fix both reduces the bandwidth use of the app.

The api querry:
/api/messageConversations/<id>?fields=read,id,name,lastUpdated,lastSenderFirstname,lastSenderSurname
Doesn't seem to have the field read any more. (atleast not for our test server).

It seems like the DHIS2 team are in the process of removing the read field from /api/messageConversations and
 /api/messageConversations/<id>. As such we needed a better method of finding out if a conversation was read by the user.

I define that as "There have been changes since you last saw it on the device".
It also serves as a flag to weather the messages in the cache for this conversation are outdated and need to be refreshed.

As such my solution gets the lastUpdated date, for exampe:"2015-12-06T21:59:27.687+0000".
Parses it to a date object and stores it in the cached object.

When a refresh happens the InboxModel.copyAttributesFrom() sets the read to false if the other date is newer. It also increments a counter in RESTSessionStorage indicating how many of the last conversations were not read.
This counter is used by onPostExecute() inside the InboxFragment asyncTask, to update the total unread messages counter and store it in the settings.

When the user clicks on the conversation to view the messages the field gets set to true.(seen)
(Also inside InboxFragment onPostExecute, if the cache was skipped, then the list is retrieved again from the cache, so that the read value is correct.)


Signed-off-by: vladisla <vladisla@student.matnat.uio.no>