Refresh merges changes.
Added a new function to CacheList.mergePageList(page, list)
that merges the changes to the cache.

Refactored the CopyAttributes to CacheListElement, since the name is more appropriate.
And added boolean getRead() to it.

The changes seem to be working, but the CacheList.setListPage() might have to be refactored in the future.
Namely keeping track of read/unread when counting number unread,
which will make it work with both using the read value from the server and using the date to extrapolate the information (workaround).

Signed-off-by: vladisla <vladisla@student.matnat.uio.no>